Windows 64-bit build support (#66)
* correcting x64 configurations to actually use x64
defining NOMINMAX to fix std::max error

* more preprocessor definitions for x64

* 64-bit libraries from the last published version of DXSDK with dinput

* should define WIN64 as well? unfortunately seem to require retention of WIN32

* added an x64 build of luaperks.lib, this compiles links and runs now

* relocate win32 luaperks.lib to src/drivers/win/lua/win32 and leave a note explaining it

* luaperks.lib x64 needed to be /MT
two more missing x64 libs
debug x64 configuration was set to compile C++ as C?

* importing the rest of the usable WIN32 preprocessor defines, debugger now functions!

* revert unnecessary VS solution version change

* 64-bit lua working (fixed the 64-bit build of luaperks.lib)

* VSUIENTRY hashes are unsigned, not signed

* more hashes that are actually unsigned

* fix a few pointer to int casts in lua console

* fix a few more pointer truncations in windows dialog stuff
a few printf arguments need a cast to int

* explicit size_t cast to quiet spurious warnings on an inline function

+++ src/lua-engine.cpp
@@ -163,16 +163,16 @@ struct LuaSaveState {
 	}
 };
 
-static void(*info_print)(int uid, const char* str);
-static void(*info_onstart)(int uid);
-static void(*info_onstop)(int uid);
-static int info_uid;
+static void(*info_print)(intptr_t uid, const char* str);
+static void(*info_onstart)(intptr_t uid);
+static void(*info_onstop)(intptr_t uid);
+static intptr_t info_uid;
 #ifdef WIN32
 extern HWND LuaConsoleHWnd;
 extern INT_PTR CALLBACK DlgLuaScriptDialog(HWND hDlg, UINT msg, WPARAM wParam, LPARAM lParam);
-extern void PrintToWindowConsole(int hDlgAsInt, const char* str);
-extern void WinLuaOnStart(int hDlgAsInt);
-extern void WinLuaOnStop(int hDlgAsInt);
+extern void PrintToWindowConsole(intptr_t hDlgAsInt, const char* str);
+extern void WinLuaOnStart(intptr_t hDlgAsInt);
+extern void WinLuaOnStop(intptr_t hDlgAsInt);
 void TaseditorDisableManualFunctionIfNeeded();
 #endif
 
@@ -6195,7 +6195,7 @@ int FCEU_LoadLuaCode(const char *filename, const char *arg) {
 	info_onstop = WinLuaOnStop;
 	if(!LuaConsoleHWnd)
 		LuaConsoleHWnd = CreateDialog(fceu_hInstance, MAKEINTRESOURCE(IDD_LUA), hAppWnd, DlgLuaScriptDialog);
-	info_uid = (int)LuaConsoleHWnd;
+	info_uid = (intptr_t)LuaConsoleHWnd;
 #else
 	info_print = NULL;
 	info_onstart = NULL;
